Modernize March server startup
